[skip ci] Add NameFormatError struct for handling name formatting errors
